added registration, preloader, views count to cms

+++ admin/includes/view_all_posts.php
@@ -2,7 +2,7 @@
 if (isset($_POST['checkBoxArray'])){
     foreach ($_POST['checkBoxArray'] as $postValueId){
         $bulkOptions = $_POST['bulkOptions'];
-        switch ($bulkOptions){
+        switch ($bulkOptions) {
             case 'publish':
                 $query = "UPDATE posts SET post_status = '{$bulkOptions}' WHERE post_id = 
 {$postValueId}";
@@ -22,6 +22,28 @@ if (isset($_POST['checkBoxArray'])){
                 $delete_status = mysqli_query($connection, $query);
                 confirmQuery($delete_status);
                 break;
+
+            case 'clone':
+                $query = "SELECT * FROM posts WHERE post_id = '{$postValueId}'";
+                $select_post_query = mysqli_query($connection, $query);
+                while ($row = mysqli_fetch_array($select_post_query)) {
+                    $post_author = $row['post_author'];
+                    $post_title = $row['post_title'];
+                    $post_category_id = $row['post_category_id'];
+                    $post_status = $row['post_status'];
+                    $post_image = $row['post_image'];
+                    $post_tags = $row['post_tags'];
+                    $post_comment_count = $row['post_comment_count'];
+                    $post_date = $row['post_date'];
+                }
+                $query  = "INSERT INTO posts(post_category_id, post_title, post_author, post_date, post_image, post_content, post_tags, post_comment_count, post_status) ";
+                $query .= "VALUES({$post_category_id}, '{$post_title}',  '{$post_author}',  now(), '{$post_image}','{$post_content}', '{$post_tags}',  '{$post_comment_count}', '{$post_status}' )";
+
+                $copy_query = mysqli_query($connection, $query);
+                if (!$copy_query){
+                    die("Query Failed " . mysqli_error($connection));
+                }
+                break;
         }
     }
 }
@@ -36,6 +58,7 @@ if (isset($_POST['checkBoxArray'])){
                     <option value="publish">Publish</option>
                     <option value="draft">Draft</option>
                     <option value="delete">Delete</option>
+                    <option value="clone">Clone</option>
                 </select>
             </div>
             <div class="col-xs-4">
@@ -53,31 +76,34 @@ if (isset($_POST['checkBoxArray'])){
         <th>Image</th>
         <th>Tags</th>
         <th>Comments</th>
-        <th>Date</th>
-        <th>View Post</th>
+        <th class="col-xs-1">Date</th>
+        <th class="col-xs-1">Post Views</th>
+        <th class="col-xs-1">View Post</th>
         <th>Edit</th>
         <th>Delete</th>
         </thead>
         <tbody>
 
         <?php
-        $query = "SELECT * FROM posts";
+        $query = "SELECT * FROM posts ORDER BY post_id DESC";
         $select_posts = mysqli_query($connection, $query);
         while ($row = mysqli_fetch_assoc($select_posts)) {
             $post_id = $row['post_id'];
             $post_author = $row['post_author'];
             $post_title = $row['post_title'];
+            //$post_content = $row['post_content'];
             $post_category_id = $row['post_category_id'];
             $post_status = $row['post_status'];
             $post_image = $row['post_image'];
             $post_tags = $row['post_tags'];
             $post_comment_count = $row['post_comment_count'];
             $post_date = $row['post_date'];
+            $post_views_count = $row['post_views_count'];
 
             echo "<tr>";
         ?>
             <td><input type="checkbox" class="checkBoxes" name="checkBoxArray[]"
-                       value="<?php echo $post_id; ?>"></td>;
+                       value="<?php echo $post_id; ?>"></td>
         <?php
             echo "<td>{$post_id}</td>";
             echo "<td>{$post_author}</td>";
@@ -95,9 +121,10 @@ if (isset($_POST['checkBoxArray'])){
             echo "<td class='col-md-2'>{$post_tags}</td>";
             echo "<td>{$post_comment_count}</td>";
             echo "<td>{$post_date}</td>";
+            echo "<td><a href='posts.php?reset={$post_id}'>{$post_views_count}</a></td>";
             echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a> </td>";
             echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a> </td>";
-            echo "<td><a href='posts.php?delete={$post_id}'>Delete</a> </td>";
+            echo "<td><a onclick=\"javascript: return confirm('Are you want to delete this post')\" href='posts.php?delete={$post_id}'>Delete</a> </td>";
             echo "</tr>";
         }
         ?>
@@ -107,9 +134,17 @@ if (isset($_POST['checkBoxArray'])){
 </form>
 
 <?php
+
 if (isset($_GET['delete'])){
     $delete_post_id = $_GET['delete'];
     $query = "DELETE FROM posts WHERE post_id = {$delete_post_id}";
     $delete_query = mysqli_query($connection, $query);
+    header("Location: posts.php");
+}
+
+if (isset($_GET['reset'])){
+    $the_post_id = $_GET['reset'];
+    $query = "UPDATE posts SET post_views_count = 0 WHERE post_id = " . mysqli_real_escape_string($connection, $_GET['reset']) . " ";
+    $reset_query = mysqli_query($connection, $query);
     header("Location: posts.php");
 }
